API BREAK: bump plugin version 6 → 8
Between the last release (12.2.1) and now, a plugin API break has occurred
(6a4d504e6539af0d86c1b5f087066bf687de695b). The plugin version has mostly not
been updated in the past, even when it probably should have been. This was
implicitly left for downstream packagers to handle. This change is an attempt to
regain rigour with respect to how plugin API breaks are communicated.

Rather than bumping to 7, this change increments by 2 to take the opportunity to
remove some legacy cruft of using the N+1 version in some situations. In both
the Autotools and CMake build systems, the motivation for these tweaks (the
`… + 1` lines removed in this change) is unknown.

The Debian Graphviz package is carrying a patch that is no longer applied, but
has some context relevant to this, versioned-plugin-config-file.diff:

  Index: b/configure.ac
  ===================================================================
  --- a/configure.ac
  +++ b/configure.ac
  @@ -67,7 +67,9 @@ GVPLUGIN_REVISION=0
   # precedence over adding, so set to 0 if both happened.
   GVPLUGIN_AGE=0

  -GVPLUGIN_CONFIG_FILE="config$GVPLUGIN_CURRENT"
  +# upstream didn't bump GVPLUGIN_CURRENT between 2.26 (gvc5) and 2.38 (gvc6).
  +# just append something to differentiate this, don't go ahead before upstream
  +GVPLUGIN_CONFIG_FILE="config$GVPLUGIN_CURRENT"a
   AC_DEFINE_UNQUOTED(GVPLUGIN_CONFIG_FILE,"$GVPLUGIN_CONFIG_FILE",[Filename for plugin configuration file.])
   AC_SUBST([GVPLUGIN_CONFIG_FILE])

This change obsoletes the patch.