core plugin: use an SVG 'viewBox' that matches the image size
When adjusting the `dpi` of an image, the unscaled dimensions would be used for
`viewBox`. When downscaling an image, the effect of this appears to be that SVG
viewers display the image at the size indicated by the `viewBox`. That is, it
appears with a significant white space border. This behavior is reportedly
counter-intuitive to users.

This change uses the scaled size for `viewBox`, removing this white space
effect. It effectively reverts 5bb60f116892261cce59d43cbbf4a3499c445adc, which
seemingly introduced the behavior prior to this commit intentionally. Quoting
the commit message:

  … It is the role of the svg viewer to correctly handle point to pixel
  conversion.

The initial comment in plugin/core/gvrender_core_svg.c perhaps also suggests
this:

  The initial <svg> element defines the SVG coordinate system so that the
  graphviz canvas (in units of points) fits the intended absolute size in
  inches. After this, the situation should be that "px" = "pt" in SVG, so we can
  dispense with stating units. Also, the input units (such as fontsize) should
  be preserved without scaling in the output SVG (as long as the graph size was
  not constrained.)

Cross-referencing the Pango plugin’s SVG output, it scales `viewBox`. So we
assume this is the correct (or at least desirable) behavior.

The page dimensions used for this are stored as integers, hence switching to
printing with `%d`. A trailing `.00` is retained to ease the transition for
downstream parsers who may be expecting explicit float values here. Though the
change may still not be completely seamless, as indicated by the edits to
reference shapes required in this change.

As an adjacent problem, db7dae6cc295755132c179ae0192c574107ef370 fixed a bug
wherein SVG `viewBox` attributes were being misinterpreted during import. As far
as I can see, the same problem exists in the core plugin (both before and after
this change) when writing out `viewBox`. Namely the third and fourth components
are incorrectly treated as corner points instead of lengths. This may need to be
fixed in future.

Gitlab: #2636, closes #2669
Reported-by: helperman