dotgen makeCompoundEdge: fix: anticipate internal cuts and no tail arrowheads
When using `rankdir=LR` an assertion failure could be triggered:

  Assertion failed: (bez->eflag), function makeCompoundEdge, file compound.c,
    line 429.

This seems to be the result of overly strict assumptions in `makeCompoundEdge`.
It assumed that (1) any cut points being operated on would lie outside the
current bounding box and (2) `ltail` would only be used in combination with
tails of beziers that have an arrowhead. To elaborate on (1):

   box
  ┌────────┐       an example of what
  │        │   ◄── `makeCompoundEdge`
  │ A──────┼─B      expected to see
  └────────┘
    A = bezier tail
    B = cut point

   box
  ┌────────┐       an example of what
  │        │   ◄── `makeCompoundEdge`
  │ A────B │          actually saw
  └────────┘
    A = bezier tail
    B = cut point

It is not clear to me whether cases like this are intended to be encountered.
But it is also not clear to me where the original assumptions stemmed from.

This change removes the assumptions, expecting that cut points may possibly lie
within the bounding box and then should be skipped. This has some interaction
with #1879 and perhaps could be seen as extending the fix there. This may not be
the last such adjustment that needs to be done to this code either.

The goal of this change was to fix `rankdir=LR`+`ltail` issues, but surprisingly
it also fixes a `splines=curved` issue.

Gitlab: #1879, fixes #1453, fixes #1949
Reported-by: Eugene Kliuchnikov
Reported-by: Peter Mueller