CI: unconditionally echo Git version in build.sh
It has recently been discovered that the Fedora ≥ 41 CI environments did not
have Git installed. While this causes no catastrophic external effects, it leads
to some internal inconsistencies. This change causes the build to fail if Git is
not installed, something we do not anticipate and want to catch.

Gitlab: closes #2672