re-enable Pylint check for 'subprocess.run' without 'check'
This was disabled in 0ce5ce75e91a3c9ed96c01faa513e6ce62d4fccb, but it seems
worthwhile being explicit with this and enforcing its use going forwards. All of
the existing uses that omitted `check` seem to have done so intentionally, but
it is easy to write new code that intends to use `check=True` and forget it.