expr scformat: avoid UB from strict aliasing violations
The originating type of the `vp` pointer passed into `scformat` varies. Thus
casting it to a `void **` and then derefencing this is Undefined Behavior. The
indirect explanation for why this is so is ISO C11 §6.5p7:

  An object shall have its stored value accessed only by an lvalue expression
  that has one of the following types:
    - a type compatible with the effective type of the object,
    - a qualified version of a type compatible with the effective type of the
      object,
    - a type that is the signed or unsigned type corresponding to the effective
      type of the object,
    - a type that is the signed or unsigned type corresponding to a qualified
      version of the effective type of the object,
    - an aggregate or union type that includes one of the aforementioned types
      among its members (including, recursively, a member of a subaggregate or
      contained union), or
    - a character type.

What exactly is a “compatible” type is explained by §6.2.7, but the short answer
is that `void *` is not compatible with other pointer types.