fix: on macOS, teach compiled test binaries how to find Graphviz libraries
In Xcode 15, Apple rewrote their linker:¹

  A new linker has been written to significantly speed up static linking.
  It’s the default for all macOS, iOS, tvOS and visionOS binaries and anyone
  using the “Mergeable Libraries” feature. The classic linker can still be
  explicitly requested using -ld64, and will be removed in a future release.
  (108915312)

A side effect of this rewrite appears to be that the new linker does not set
rpaths for libraries that it links in. I do not know if manually setting the
rpath like this is the right solution, but it is the only thing I have found
that fixes the macOS+CMake CI job. The macOS+Autotools job remains broken for
unknown reasons.

¹ https://developer.apple.com/documentation/xcode-release-notes/xcode-15-release-notes

Gitlab: #2667