expr: support customizing the array length operator ('#')
This is a step towards supporting `#ARGV`. Though `ARGV` is an array, libexpr is
not actually aware of it. It is a global defined by libgvpr, outside of the
context passed to libexpr to evaluate. Because of the dependency direction
(libgvpr depends on libexpr, rather than the other way around), libexpr cannot
simply check for or reference `ARGV`. Instead we need to thread accesses like
`#ARGV` through callbacks into libgvpr.

Gitlab: #2582