expr: support customizing the array membership operator ('in')
This is a step towards supporting `foo in ARGV`. Though `ARGV` is an array,
libexpr is not actually aware of it. It is a global defined by libgvpr, outside
of the context passed to libexpr to evaluate. Because of the dependency
direction (libgvpr depends on libexpr, rather than the other way around),
libexpr cannot simply check for or reference `ARGV`. Instead we need to thread
accesses like `foo in ARGV` through callbacks into libgvpr.

A couple of subtleties here:
  • The custom handler is branched to in `evaldyn` so that the user’s
    implementation does not need to resolve/evaluate the LHS. This is done
    within `evaldyn` so the user’s implementation does not need to care about
    context etc.
  • The result of `foo in bar` is considered to be `0` if the user chooses not
    to install a custom handler for this situation. This seems the most
    logically coherent choice.

Gitlab: #2582