util gv_strndup: avoid UB call to 'memcpy'
Calling `memcpy` with either source or destination as `NULL` pointers, even with
length 0, is Undefined Behavior. This is a latent issue in that there are no
current calls into `gv_strndup` that pass `NULL`. But it seems wise to avoid
this unnecessary sharp edge.

Gitlab: https://gitlab.com/graphviz/graphviz/-/issues/2678#note_2473116388