tcldot: fix recognition of '$g layout'
The `layout` command on graphs was incorrectly implemented twice, once as
`layoutnodes` and once as `layoutedges`. Surprisingly this mistake has existed
since the first revision of Graphviz. However, it was masked by a further bug:
the comparison against known commands was using the _user string’s_ length. This
resulted in matching any prefix. E.g. `$g lay` would be incorrectly recognized
as `$g layoutedges`.

080346bede08ae53f54c7243ad685691c466414f fixed this second bug, unmasking the
first. From the commit message and surrounding context, it seems the functional
effect of this change was unintended; it was believed this change was a
simplification with no functional impact.

Gitlab: fixes #2618
Reported-by: Gregor