CI: remove pre-installed Python in Cygwin+CMake job
The Cygwin+CMake job has begun failing:

  [ 13%] generate color table
  /cygdrive/c/GitLab-Runner/builds/smattr/graphviz/lib/common/htmlparse.y: warning: 2 shift/reduce conflicts [-Wconflicts-sr]
  /cygdrive/c/GitLab-Runner/builds/smattr/graphviz/lib/common/htmlparse.y: note: rerun with option '-Wcounterexamples' to generate conflict counterexamples
  C:\Python313\python.exe: can't open file 'C:\\cygdrive\\c\\GitLab-Runner\\builds\\smattr\\graphviz\\lib\\common\\make_colortbl.py': [Errno 2] No such file or directory

The cause of this appears to be that Python was added to the Windows base image¹
and now CMake detects this Python 3.13 in preference to the Cygwin-installed
Python 3.9. This Windows Python does not understand how to interpret or access
Cygwin paths.

In theory there are ways to control CMake’s `find_package(Python)` to direct
which Python it selects.² In practice, none of these seem to work.³ Bafflingly,
uninstalling the pre-installed Python does not work either.
`choco uninstall python3` claims to succeed, but then CMake is still able to
find and use this Python that was supposedly uninstalled.

This change instead takes a more extreme approach and forcefully deletes the
pre-installed Python from disk. CMake now detects only the remaining
Cygwin-installed Python.

¹ https://gitlab.com/gitlab-org/ci-cd/shared-runners/images/gcp/windows-containers/-/merge_requests/72
² https://cmake.org/cmake/help/latest/module/FindPython.html#hints
³ https://gitlab.kitware.com/cmake/cmake/-/issues/21186,
  https://gitlab.kitware.com/cmake/cmake/-/issues/19492

Gitlab: fixes #2673
Gitlab: https://gitlab.com/gitlab-org/ci-cd/shared-runners/images/gcp/windows-containers/-/issues/14