CI: fix incorrect rewriting of macOS plugin→plugin links
The plugins are installed to a subdirectory of where the libraries are
installed. So the first branch of this loop that was checking for library
prefixes was incorrectly also matching plugin links. E.g.:

  + otool -L …/lib/graphviz/myplugin.dylib
  …
  …/lib/graphviz/anotherplugin.dylib
  …
  + install_name_tool -change …/lib/graphviz/anotherplugin.dylib \
    @loader_path/../graphviz/anotherplugin.dylib \ ◄────────────────┐
    …/lib/graphviz/myplugin.dylib                                   │
                                                                    │
                                      Incorrect. This should have been rewritten
                                      to simply @loader_path/anotherplugin.dylib.

We fix this by swapping the order of the two branches, so the more specific one
identifying plugins takes precedence.

This problem is latent because (1) no plugin links to another plugin and (2) the
plugins’ self-links (a dynamic library’s identity of where it itself lives on
disk) are incorrectly not being affected by this right now.

Gitlab: #2667