CI: fix relocation adjustments of macOS compiled artifacts
In Xcode 15, Apple rewrote their linker:¹

  A new linker has been written to significantly speed up static linking.
  It’s the default for all macOS, iOS, tvOS and visionOS binaries and anyone
  using the “Mergeable Libraries” feature. The classic linker can still be
  explicitly requested using -ld64, and will be removed in a future release.
  (108915312)

A side effect of this rewrite appears to be that the new linker does not set
rpaths for libraries that it links in. 17918445def008c934839983b887c6b4a0eb069a
fixes this issue for the macOS CMake job, while the present change fixes it for
the macOS Autotools job.

I only have partial knowledge of how these linking details work on macOS, so the
following is somewhat an educated guess. There were a couple of problems with
the relocation script:
  1. The first item listed by `otool -L …` is not a dependent but rather the
     “install name” of the library itself. This cannot be changed by
     `install_name_tool -change …` but instead needs to be changed with
     `install_name_tool -id …`.
  2. Even with this adjustment, the relocation script would have remained
     incorrect because, in the context of the install name, `@loader_path`
     refers to the executable/library through which the current library is being
     loaded, rather than the loadee here. So there is no obvious self-contained
     way to rewrite this reference.

This change attempts to resolve this by (1) selecting an rpath of
<install prefix>/lib, (2) applying this to the Graphviz binaries, and (3) using
paths relative to this for libraries’ install names. This cooperates with the
changes from 17918445def008c934839983b887c6b4a0eb069a to make the libraries
findable by binaries compiled by the test suite which are given the same rpath.

The problems above were challenging to diagnose because most of the Graphviz
maintainers do not have Apple hardware. Debugging had to be done through CI
runners which is a little like seeing through a scanner darkly. Further, Apple
only seem to publish man pages through Xcode, not online,² making surprises like
e.g. `install_name_tool -change …` silently failing when asked to rewrite the
install name hard to discover.

Because Apple does not officially document how the new linker works, one has to
use third-party sources which are frequently out-of-date or inaccurate. E.g.³

  What happens when the `-id` name doesn't match the `.dylibs`          ┐
  location on disk? Nothing. …                                          │
                                                                        ├  bad
  Why have an `-id` name at all? Surely if the linker is told to look   │ advice
  for libraries in `dirA`, `dirB`, `dirC` (using `-L` for example) and  │
  it finds the library it's looking for then why does the library       │
  itself have to have its `-id` stamped into it? No idea; it's          │
  probably some archaic nonsense.                                       ┘

In reality, the new linker seems to read the install name during link time and
then apply this decision at runtime. That is, if you have foo/bar.dylib with an
install name of baz/bar.dylib, you can link with something like
`cc … -Lfoo -lbar` but at runtime your binary will try to load baz/bar.dylib. I
assume this is actually a deliberate “feature” allowing something akin to
Windows’ “import libraries,” small stubs that refer to a full dynamic library
elsewhere or to be provided later.

Some further materials on these topics which may be enlightening to any readers
who made it this far:
  • https://github.com/spack/spack/issues/26552
  • https://github.com/spack/spack/pull/26608
  • https://stackoverflow.com/questions/9263256/why-is-install-name-tool-and-otool-necessary-for-mach-o-libraries-in-mac-os-x
  • https://itwenty.me/posts/01-understanding-rpath/
  • https://matthew-brett.github.io/docosx/mac_runtime_link.html
  • https://developer.apple.com/forums/thread/736719

¹ https://developer.apple.com/documentation/xcode-release-notes/xcode-15-release-notes
² https://developer.apple.com/documentation/os/reading-unix-manual-pages
³ https://stackoverflow.com/questions/35220111/install-name-tool-difference-between-change-and-id

Gitlab: fixes #2667