gvpr: fix: when setting subgraph defaults, explicit-ize root graph defaults
When setting a subgraph attribute default, the root graph would first be
consulted to see if it had a registered default. If not, the new default was
considered something that should be applied to the root graph. It is unclear to
me why the logic was designed this way and it seems this behavior is counter to
what users expect. This change prevents this “hoisting” of defaults to the root
graph by forcing the root graph’s implicit default to first become explicit in
this scenario.

Gitlab: fixes #2639
Gitlab: https://gitlab.com/graphviz/graphviz/-/issues/2578#note_2018578508
Reported-by: Steve Roush