CI: increase timeout for Windows ASan jobs 1h → 2h
These jobs often exceed 1h. Timeouts have not typically not proved a problem for
maintainers because the project-level timeout for the Graphviz repository is set
to 3h. But in contributors’ forks, their default of 1h takes precedence.

According to the docs:¹

  The job-level timeout can be longer than the project-level timeout, but can’t
  be longer than the runner’s timeout.

So this value of 2h is legal (greater than the default contributors’ forks’
project-level timeout of 1h) and also the maximum we can set (equal to the
Windows shared runner timeout).

¹ https://docs.gitlab.com/ci/yaml/#timeout

Gitlab: closes #2662
Suggested-by: Mark Hansen <mark@markhansen.co.nz>