CI: expire website artifacts within 1 week
The `pages:deploy` CI job has begun failing with a tooltip error:

  artifacts for pages are too large: 1051961349

It is unclear to me what this is referring to. Viewing artifacts in the Gitlab
UI, the artifacts for the last successful run of this job are indeed large
(234.61MB) but nowhere near this claimed ~1GB limit. It is possible this is
referring to the uncompressed size of the artifacts, ~1.2GB.

This change makes a guess that the error is referring to the accumulated size of
retained artifacts for this job and reduces their retention period. Prior to
this change, it was 30 days.¹ It is unlikely anything more than the last
successful run of this job is required to be retained, but 1 week is aligned
with other CI jobs limited in this way.

While making this change, it was observed that many CI jobs export artifacts
with no retention period defined and/or, more puzzlingly, export artifacts
“always” (regardless of success or failure of the job). This has been left as-is
for now.

¹ https://docs.gitlab.com/administration/settings/continuous_integration/#set-default-artifacts-expiration

Gitlab: #2684