sfdpgen SparseStressMajorizationSmoother_new: simplify check for zeroed array
This code seemed to be using `x[i] * x[i]` as a substitute for taking the
absolute value of entries, to rule out negative values canceling out positive
values. Neither this nor the summing are necessary. We can simplify the work
being done here to both avoid the expense of the sum and bail out as soon as we
know the answer.

Zooming out, what this code is doing is fairly questionable IMHO. The array `x`,
in at least one path leading into this function, is generated by calls to
`drand()`. These may coincidentally all return 0, but it is also possible for
the re-initialization here to also all return 0. That is, there is no guarantee
after this code block that `x` contains a non-zero value. Nevertheless this is
left as-is.