common initMapData: fix: also indicate change when ID but not URL updated
Address Sanitizer reports about `test_2619_1[2619_2.dot-pdf-2619_2]`:

  Direct leak of 32 byte(s) in 4 object(s) allocated from:
      #0 0x7ffff74b4887 in __interceptor_malloc ../../../../src/libsanitizer/asan/asan_malloc_linux.cpp:145
      #1 0x7ffff7456b21 in __interceptor_strndup ../../../../src/libsanitizer/sanitizer_common/sanitizer_common_interceptors.inc:414
      #2 0x7ffff6b51de1  (/tmp/tmp.OI4zOPzZOH/lib+0x551de1)
      #3 0x7ffff6b53a1d  (/tmp/tmp.OI4zOPzZOH/lib+0x553a1d)
      #4 0x7ffff6b58eb3  (/tmp/tmp.OI4zOPzZOH/lib+0x558eb3)
      #5 0x7ffff6b58f77  (/tmp/tmp.OI4zOPzZOH/lib+0x558f77)
      #6 0x7ffff6c1c8af  (/tmp/tmp.OI4zOPzZOH/lib+0x61c8af)
      #7 0x7ffff6ac9ece  (/tmp/tmp.OI4zOPzZOH/lib+0x4c9ece)
      #8 0x7ffff6ace40d  (/tmp/tmp.OI4zOPzZOH/lib+0x4ce40d)
      #9 0x7ffff6ad2e2f  (/tmp/tmp.OI4zOPzZOH/lib+0x4d2e2f)
      #10 0x7ffff6b56a41  (/tmp/tmp.OI4zOPzZOH/lib+0x556a41)
      #11 0x7ffff6bcb71d  (/tmp/tmp.OI4zOPzZOH/lib+0x5cb71d)
      #12 0x7ffff6c4254f  (/tmp/tmp.OI4zOPzZOH/lib+0x64254f)
      #13 0x7ffff6c6000f  (/tmp/tmp.OI4zOPzZOH/lib+0x66000f)
      #14 0x7ffff6c61be1  (/tmp/tmp.OI4zOPzZOH/lib+0x661be1)
      #15 0x7ffff6c631bd  (/tmp/tmp.OI4zOPzZOH/lib+0x6631bd)
      #16 0x7ffff6c6ba0f  (/tmp/tmp.OI4zOPzZOH/lib+0x66ba0f)
      #17 0x5555555555f8 in main cmd/dot/dot.c:73
      #18 0x7ffff5a29d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58

Working around #2526, we can learn the leak originates from allocations for ID
strings when processing an HTML table. The root cause is that the “has anything
changed?” indication `initMapData` was answering “no” if the ID had changed but
the URL had not. While this may not lead to visual incongruities, a problem was
that the caller did not realize ID had been reallocated and needed to be
restored to its old value.

Philosophically, it seems the right thing to answer “yes” if the ID alone
changes. While this may not lead to visual differences, for cross-references to
be correct code must have an accurate idea of the IDs of each entity.

Gitlab: #2526
Reported-by: GCC 11.4.0 -fsanitize=address