mingle: fix use-after-free of 'ANNkd_tree's
c4bde8f60fd6de34f2ce31aea514b3580225303a, in fixing a memory leak, introduced a
use-after-free. The underlying problem is that the `ANNkd_tree` class has raw
pointer members but no user-defined move operator. The default move operator
does something undesirable, duplicating a heap pointer. Replicating something
like the situation in C:

   1: void ctor(ANNkd_tree *self) {
   2:   self->x = malloc(10);
   3: }
   4:
   5: void dtor(ANNkd_tree *self) {
   6:   free(self->x);
   7: }
   8:
   9: void move(ANNkd_tree *dst, ANNkd_tree *src) {
  10:   dst->x = src->x;
  11: }
  12:
  13: int main(void) {                    // in C++:
  14:   ANNkd_tree a; ctor(&a);           // ANNkd_tree kdTree(…);
  15:   ANNkd_tree b; ctor(&b);           //          ANNkd_tree(…);
  16:   dtor(&a); move(&a, &b); dtor(&b); // kdTree = ─┘
  17:
  18:   dtor(&a);
  19:   return 0;
  20: }

The problem arises on line 16. The `move` duplicates the pointer `b->x` into
`a->x`. But then the destructor is called on `b` (the intermediate rvalue on the
RHS in the originating C++). After this, use of `a->x` is referencing freed
memory. Note that, even if `a->x` were never referenced`, this code would still
be buggy as the eventual destruction of `a` causes a double-free.

Arguably this is a bug in libANN. But this library appears to be unmaintained.
It was developed prior to C++11 and concepts like “the rule of three/five/zero”¹
were widely established, so sharp edges like this in C++ code from this era are
common.

¹ https://en.cppreference.com/w/cpp/language/rule_of_three

Reported-by: Phormio
Gitlab: fixes #2599