ingraphs dflt_read: make use of the 'filename' provided
This change essentially inlines the `agread` call and then threads the
`filename` information through the result. This has no functional effect for
users because the calling ingraphs code was already calling `agsetfile` prior to
this with the same filename. It will take effect in upcoming changes to remove
`agsetfile`.

Gitlab: #2558