API BREAK: cgraph: remove 'agsetfile'
This function manipulated global state which was proving a barrier to
thread-safety. The recent changes to `agconcat` make it possible for callers to
specify filename in a thread-safe way. Debian Code Search finds no third-party
uses of this function¹ so its removal should not be too disruptive. This change
does the removal suggested in 58f13887618e5a0c1ce77216a76c5a9cb58bb138.

¹ https://codesearch.debian.net/search?q=agsetfile&literal=1

Gitlab: #2558
Suggested-by: Bjarne von Horn <vh@igh.de>