API BREAK: cgraph: track which graph was used to allocate an 'Agsym_t'’s strings
Currently freeing `Agsym_t`s relies on the `Ag_G_global` global to propagate
information about which graph owns the symbol. This is proving a barrier to
thread-safety, as concurrent attempts to operate on symbols interfere with each
other’s updates to `Ag_G_global`. This change is a step towards removing this
global altogether. Note that this extra field added to `Agsym_t` is not yet
used.

Gitlab: #2558, !4147
Suggested-by: Bjarne von Horn <vh@igh.de>