cgraph setattr: fix allocation home of global attributes
This code was creating a new symbol, containing `name` and `defval` allocated
out of `g`’s string pool. It then goes on to add the new symbol to `rdict`, a
dictionary not of `g` but of the root graph of `g`.

This was a latent bug, in that it causes no problems currently because all
objects involved (the symbols, subgraph(s), and the root graph) are all torn
down at the same time. Deallocation of pooled strings happens using the
`Ag_G_global`, something that always points at the current graph during
deletion. That is, deallocation of a subgraph leaves dangling pointers
(`Agsym_t.owner`) in the root graph’s attributes that still point to this freed
memory, however these pointers are not currently referenced during the
subsequent deallocation of the root graph and its attributes. During
deallocation of the root graph’s attributes, `Ag_G_global` points at the root
graph itself, masking this bug.

The bug was discovered while altering `freesym` to deallocate pooled strings
through the `Agsym_t.owner` pointer instead of the `Ag_G_global` pointer (an
upcoming commit). After this change, deallocating symbols in the root graph no
longer tried to free these strings through `Ag_G_global`, but instead tried to
free them through a pointer to the (already deallocated) subgraph. Naturally
this caused significant misbehavior and crashes, thankfully diagnosed readably
by Address Sanitizer.

Surprisingly this bug has been present from the very first revision in version
controlled Graphviz history.

Note that a similar but reversed situation exists in `unviewsubgraphsattr`: an
attribute has its strings allocated out of the root graph’s string pool and the
attribute is then added to a subgraph’s dictionary. In theory, this kind of
mismatch should be safe because a root graph can never be deallocated while one
of its subgraphs is still live. ff4c6ce29a796f11f70235c01b744929c6dfb0a9 that
added this code implies it is necessary to handle the creation of new root
graphs.

Gitlab: #2558, !4147