util: bring back integer-to-string converter
The previous `itos` implementation was removed in
edea92223cadc27613b9955e865bdfc3dd684102 because it was unsafe under C99 and it
was estimated that migrating to C11 was unlikely to happen soon. Three years
later, MSVC supports C11 and Graphviz is compiled as C11. Thus it is safe to
reinstate this functionality.

In comparison to the previous version, this one is widened to also handle `long`
and `long long` values.

Understanding why usage of this macro is safe requires some subtle analysis of
the C11 (and optionally C++17) standard. Discussion comments on the following
page are informative:
  • https://wiki.sei.cmu.edu/confluence/display/c/EXP35-C.+Do+not+modify+objects+with+temporary+lifetime

----

`ITOS` has one main sharp edge. As discussed in its comments, incorrect use can
lead to a dangling pointer:

  char *p = ITOS(42);
  // p is now a dangling pointer, pointing to deallocated stack memory

While runtime instrumentation like Address Sanitizer can catch this,
contemporary compilers cannot spot this issue at compile-time. It is possible to
make this safer by using a C99 compound literal in `ITOS`’s definition:

  static inline char *itos_(char *s, long long i) {
    sprintf(s, "%lld", i);
    return s;
  }

  #define ITOS(i) itos_((char[41]){0}, (i))

This makes the former usage of `p` safe because compound literals have block
lifetime. However alternative forms still result in dangling pointers:

  char *p;
  {
    p = ITOS(42);
  } // ← p is now a dangling pointer, pointing to deallocated stack memory

Given that this does not fully solve the problem and is incompatible with C++,¹
the choice has been made to retain the current simpler formulation and the
advice to simply never store the result of `ITOS`. The sharp edge seems worth it
for the simplification `ITOS` provides in calling code. In future, perhaps a
safer way to define this macro will be found.

Further reading on this topic:
  • https://en.cppreference.com/w/c/language/compound_literal.html
  • https://www.drdobbs.com/the-new-c-compound-literals/184401404
  • https://gcc.gnu.org/onlinedocs/gcc/Compound-Literals.html

¹ Compound literals are not supported in ISO C++. Some compilers (e.g. GCC)
  support compound literals in C++ as an extension. However, in GCC’s case,
  their lifetime is the containing full expression _not_ the containing block,
  differing from C. This discrepancy makes them difficult to use safely in mixed
  mode code.

Gitlab: #2229, #2535
Gitlab: https://gitlab.com/graphviz/graphviz/-/merge_requests/2575#note_907097811