CMake: forward port 'exec_program' to 'execute_process'
`exec_program` has been deprecated since CMake 3.0:¹

  Changed in version 3.28: This command is available only if policy `CMP0153` is
  not set to `NEW`. Port projects to the `execute_process()` command.

  Deprecated since version 3.0: Use the `execute_process()` command instead.

These changes to forward port this code are based on commit
ca42c0c540313b55895a4a153d9b78c2d58e4fa7 to the CMake community wiki¹ that
updated the FAQ entry on which this code was originally based. The unused
`OUTPUT_VARIABLE` is dropped, the `COMMAND` line is made explicit, and the
`rm_retval` comparison is done on integers. These deviations from the example
are improvements and/or simplifications.

¹ https://gitlab.kitware.com/cmake/community/-/wikis/home