dotgen edgecmp: remove reuse of 't0', 't1'
It seems the original implementation of this function reused `t0` and `t1`
purely for convenience. 113799f20714a420dde7b179c8238eecfbbb10ff then altered
their type to `double`, overlooking that this now made this section of the code
coerce an integer difference into a `double` only to force it back into an
`int`. The “ugly, but explicit as to how we avoid equality tests on fp numbers”
never needed to be a comparison on floating-point numbers at all. This change
undoes this noise, converting to more readable and intuitive code.