dotgen make_vn_slot: fix: reallocate rank list through owning pointer
The rank structure, `rank_t`, contains two pointers to node lists, `v` and `av`.
If you believe their comments, `av` is to point to the underlying allocation.
`make_vn_slot` was reallocating the node list through `v`, the _unowning_
pointer of these two lists. A side effect of this was that `av` remained
pointing at the (now potentially freed) old list.

I say “if you believe” in the preceding paragraph because the ownership
semantics around these node lists is very murky. `v` is regularly reassigned and
`av` entries are sometimes accessed with an index that appears out-of-bounds.
The number of entries `av` is allocated/re-allocated with rarely corresponds to
anything obvious in the surrounding code. There is also the (appropriately
named?) function `abomination`, that appears to do all manner of
memory-safety-violating actions. Nevertheless the bug fixed in this commit is
one of the few issues ASan diagnoses with these actions.

The assertion included in this commit is an expression of something I believe to
be true at this point, but I am not 100% sure. If this assertion is false,
re-allocating through `av` here is not valid.

Two surprises were encountered while debugging this:
  1. Address Sanitizer seemed unable to diagnose this as a double `free`. It
     instead called this “attempting free on address which was not malloc()-ed”.
  2. Running the #2331 test case’s DOT source through `dot` itself did not
     trigger this failure.

I believe what is happening here is that triggering this issue is dependent on
the shape of the heap at the time of `make_vn_slot`’s re-allocation.
Simplifying:

  node_t **av = calloc(2, sizeof(node_t *));
  node_t **v = av;
  …
  v = realloc(v, 3 * sizeof(node_t *));

Whether the `realloc` actually leads to moving the underlying allocation is
dependent on the surrounding heap layout. Two scenarios:

     ┌───┬────┬───────────────┐         ┌───┬──────┬─────────────┐
  A  │ … │ av │╳╳╳╳╳╳╳╳╳╳╳╳╳╳╳│    →    │ … │ av   │╳╳╳╳╳╳╳╳╳╳╳╳╳│
     └───┴────┴───────────────┘         └───┴──────┴─────────────┘
                                realloc
     ┌───┬────┬──────┬────────┐         ┌───┬────┬──────┬──────┬─┐
  B  │ … │ av │ foo  │╳╳╳╳╳╳╳╳│    →    │ … │╳╳╳╳│ foo  │ av   │╳│
     └───┴────┴──────┴────────┘         └───┴────┴──────┴──────┴─┘

  ╳ = unallocated heap space

In scenario A, there is enough room to expand the allocation in place. So the
original `av` pointer remains valid after the `realloc` call. In scenario B,
`av` cannot be expanded in place because there is another live allocation
following it. `av` instead needs to be moved, leaving the old `av` pointer
dangling. Which of these scenarios you hit depends on a number of factors
including non-deterministic elements like Address Space Layout Randomization.

I believe these scenarios explain (1) and (2) above. (1) hits scenario B and
thus has discarded any tracking information about the original `av` allocation
and knows it only as “something not heap-allocated.” It seems ASan does not
track heap pointers that are moved through `realloc` as being semantically
freed. (2) hits scenario A and thus sees no problems with memory accesses, even
through the old `av` pointer.

Gitlab: fixes #1494, fixes #2331
Reported-by: Google Autofuzz project
Reported-by: Gordon Smith