fix: set '\N' default node label during parsing if not already set
`gvContext` and `gvContextPlugin` set the default node label to `"\N"`.
Manipulating a global at this point¹ has the side effect that other operations
that depend on the node label default differ in their outcome depending on
whether they are run before or after a context is created. Notably, this
includes graph parsing.

This change duplicates the default setting into the common function for graph
parsing. This is still not good practice (setting a global when parsing a graph
is an issue for thread-safety), but is at least less surprising to users.

A side effect of this change is that GVPR output now differs. However this
difference is arguably a fix too, in that GVPR output now more closely matches
dot canonicalization.

¹ It is unclear to me why it is appropriate to set this kind of global state
  when constructing a new context. Commit messages like
  ea7adf22d0aa3aaa5a5aa246ebeef289ee563101 suggest the original authors did not
  know either.

Gitlab: #2558, fixes #2434, fixes #2436
Reported-by: Chiel ten Brinke
Reported-by: Podbrushkin