dotgen: fix: make adjacency matrices sparse
When computing an adjacency matrix, a square matrix with side lengths of the
number of ranks would be allocated. This seems vaguely reasonable, yet the
example of #1472 shows it is not. This example caused out-of-bounds accesses
past the end of the computed adjacency matrix.

Rather than attempt to derive a corrected upper bound for the dimensions of this
matrix, this change instead makes the row count dynamic. While the number of
columns is fixed, accessing rows past the expected upper bound now causes the
matrix to be dynamically expanded. This is most likely not fully correct (if we
have the expected area of the matrix wrong, it seems unlikely that the expected
number of columns of the matrix is right), yet it is unclear what better
alternative exists.

Gitlab: fixes #1472
Reported-by: Google Autofuzz project