dotgen interclexp: fix: skip creating flat edges that duplicate virtual edges
When creating flat edges to correspond to inter-cluster edges in the input
graph, only previously existing flat edges would be checked for. An
unanticipated edge case occurred when a previously existing flat edge had been
deleted and readded as a reversed virtual edge. In this scenario, this virtual
edge would be encountered, not recognized as originating from a flat edge, and
then used to recreate a new flat edge. The visible result to end users was that
any such edges would be unexpectedly duplicated in their output image. This bug
seems to have existed ~forever.

I have an incomplete understanding of this code and what it is trying to
achieve. The previous paragraph is based on my best guess and some dynamic
debugging of the #2615 test case. It is possible this fix is incomplete or
incorrect.

Gitlab: fixes #162, fixes #258, fixes #2615
Mantis: #2455, #2550
Reported-by: DelTree
Reported-by: Łukasz Nojek
Reported-by: truss