common routesplines_: fix inability to process horizontal/vertical splines
When a set of spline points happened to describe a horizontal or vertical line,
the loop attempting to shrink bounding boxes would never converge. The root
cause seems to be simply that this algorithm never anticipated such splines. The
scaling it uses to attempt to bound spline points within the boxes repeatedly
derives a 0 scaling factor, and hence does not make progress. Thankfully such
splines, if the algorithm were to converge, have an obvious outcome that we can
just apply instead.

Gitlab: fixes #241
Mantis: #2472
Reported-by: perlindgren
Reported-by: Steve Roush