rtest: fix: expect to find reference files in our own directory
As discussed in prior commits, this fixes a bug in
0972f09531e16352fcd49f5a6740552cf9d4a57a wherein the change in working directory
was not accounted for.

Although this un-xfails these test cases they still do not actually pass.
However the way in which they fail is not visible to Pytest. This will be
addressed in an upcoming commit.

Gitlab: #2197
Gitlab: https://gitlab.com/graphviz/graphviz/-/merge_requests/4348#note_2592297747