common gvRenderJobs: repair debug compilation under GCC 14
GCC 14 gets more strict and turns some things that were previously warnings into
errors:

  emit.c: In function 'gvRenderJobs':
  emit.c:3973:35: error: assignment to 'const char **' from incompatible pointer
    type 'char **' [-Wincompatible-pointer-types]
   3973            job->common->show_boxes = show_boxes_front(&Show_boxes);

This change effectively reverts 180f77a87f216e9b60b3eb1d5acce11cb5d6c034 which
introduced this issue. I am not sure why
180f77a87f216e9b60b3eb1d5acce11cb5d6c034’s commit message says `Show_boxes` was
a `const char **`. It was not; it was a `char **`.

A cast is not a particularly clean way of solving this and we should just
realign these types at the next opportunity for an API break.

Gitlab: #2699
Reported-by: Steve Roush
Reported-by: GCC 14.2.0 -Wincompatible-pointer-types